Fix clocks_remaining still surfacing as 4.0-style floats
Caught by Scott: "3.7 'charges' 4.0" -- seconds_used was correctly an int
per the last commit, but BASE_CLOCK was still 25.0 and the bonus was still
+5.0/+0.0, so clocks[pid] (int base minus int charges, starting from a
float) stayed a float throughout, even though every value in it was
numerically whole. clocks_remaining (DuelResult) and the per-turn
clock_remaining fed off that same float accumulator.

BASE_CLOCK is now 25 (int), the bonus is +5/+0 (int), DuelResult.clocks_remaining
is typed Dict[int, int], and the per-turn clock_remaining clamp uses max(0, ...)
instead of max(0.0, ...). Verified directly: clocks_remaining now returns
{0: 0, 1: 25} with real int values, not 0.0/25.0, and a 30-show regression
confirms every seconds_used, clock_remaining, and clocks_remaining value in
the event stream is a proper int.

+++ prototype/engine/duel.py
@@ -65,7 +65,12 @@ from .models import Player
 
 
 QUESTION_CAP = 30  # total across BOTH players in one duel, not per-player
-BASE_CLOCK = 25.0
+# An int, not 25.0 -- every charge against this clock (attempt.seconds_used)
+# is now a whole number by construction (Scott's rule), so keeping the
+# clock itself an int end to end means clocks_remaining never drifts back
+# into "4.0"-looking float output despite every individual charge already
+# being a clean integer.
+BASE_CLOCK = 25
 # A live model's "pass" is entirely voluntary -- attempt_question only
 # registers one if the model's reply literally starts with "PASS" (see
 # ollama_agent.py). A model that never volunteers that word can otherwise
@@ -84,7 +89,7 @@ class DuelResult:
     loser_id: int
     reason: str                      # "timeout" or "question_cap"
     turns_log: List[Dict[str, Any]] = field(default_factory=list)
-    clocks_remaining: Dict[int, float] = field(default_factory=dict)
+    clocks_remaining: Dict[int, int] = field(default_factory=dict)
     questions_seen: Dict[int, int] = field(default_factory=dict)
 
 
@@ -118,15 +123,15 @@ def run_duel(challenger: Player, defender: Player, domain: Domain,
              agents: Dict[int, ScriptedAgent], rng: random.Random,
              challenger_bonus: bool = False, defender_bonus: bool = False,
              used_questions: Optional[Set[Tuple[str, str]]] = None,
-             base_clock: float = BASE_CLOCK, question_cap: int = QUESTION_CAP,
+             base_clock: int = BASE_CLOCK, question_cap: int = QUESTION_CAP,
              on_turn: Optional[Callable[[Dict[str, Any]], None]] = None) -> DuelResult:
 
     if used_questions is None:
         used_questions = set()
 
     clocks = {
-        challenger.id: base_clock + (5.0 if challenger_bonus else 0.0),
-        defender.id: base_clock + (5.0 if defender_bonus else 0.0),
+        challenger.id: base_clock + (5 if challenger_bonus else 0),
+        defender.id: base_clock + (5 if defender_bonus else 0),
     }
     seen = {challenger.id: 0, defender.id: 0}
     turns_log: List[Dict[str, Any]] = []
@@ -176,15 +181,16 @@ def run_duel(challenger: Player, defender: Player, domain: Domain,
             "answer": question.answer, "guess": attempt.guess, "outcome": attempt.outcome,
             "correct": attempt.correct, "seconds_used": attempt.seconds_used,
             "distractors": distractors, "live": attempt.live,
-            # Clamped to 0 for display: the real clock can dip slightly
-            # negative the instant it crosses zero, but the audience should
-            # never see a negative number on screen. Elimination logic
-            # below still uses the real, unclamped clocks[pid] value. A
-            # clean int, not a rounded float -- every seconds_used charged
-            # against it is already a whole number (Scott's rule), so the
-            # running total is always whole too; no fractional remainder
-            # is possible to round away here.
-            "clock_remaining": int(max(0.0, clocks[pid])),
+            # Clamped to 0 for display: the real clock can dip negative the
+            # instant it crosses zero (an int charge subtracted from an int
+            # clock, so still a clean negative int, e.g. -2), but the
+            # audience should never see a negative number on screen.
+            # Elimination logic below still uses the real, unclamped
+            # clocks[pid] value. base_clock and every seconds_used charged
+            # against it are ints by construction (Scott's rule), so this
+            # was always a whole number -- max(0, ...) just clamps the
+            # sign, nothing here rounds or truncates a fraction away.
+            "clock_remaining": max(0, clocks[pid]),
         })
         # Fire as each turn is actually computed, not just appended to
         # turns_log for the caller to replay after the whole duel finishes.
